Refactor: 검색페이지 컴포넌트 분할 및 hooks 생성
