+++ src/store/searchSelectorAtom.ts
@@ -1,0 +1,16 @@
+import { atom } from "recoil";
+
+export const searchStateAtom = atom({
+  key: "searchState",
+  default: {
+    selectedRegion: 99,
+    selectedType: 99,
+    selectedOptions: [99],
+    isRegionSelected: false,
+    isRegionHovered: false,
+    isTypeSelected: false,
+    isTypeHovered: false,
+    isOptionsSelected: false,
+    isOptionsHovered: false,
+  },
+});
