+++ src/store/searchSelectorData.ts
@@ -1,0 +1,31 @@
+export const regions = [
+  { name: "전국", value: 99 },
+  { name: "서울시", value: 0 },
+  { name: "경기도", value: 1 },
+  { name: "강원도", value: 2 },
+  { name: "충청도", value: 3 },
+  { name: "전라도", value: 4 },
+  { name: "경상도", value: 5 },
+  { name: "제주도", value: 6 },
+];
+
+export const types = [
+  { name: "전체 타입", value: 99 },
+  { name: "호텔", value: 0 },
+  { name: "콘도", value: 1 },
+  { name: "호스텔", value: 2 },
+  { name: "펜션", value: 3 },
+  { name: "모텔", value: 4 },
+  { name: "민박", value: 5 },
+  { name: "게스트하우스", value: 6 },
+  { name: "홈스테이", value: 7 },
+  { name: "레지던스", value: 8 },
+  { name: "한옥", value: 9 },
+];
+
+export const options = [
+  { name: "상관 없음", value: 99 },
+  { name: "주차 가능", value: 0 },
+  { name: "조리 가능", value: 1 },
+  { name: "픽업 가능", value: 2 },
+];
